fix：年次休暇管理时间修正

+++ src/main/java/jp/smartcompany/job/modules/tmg/paidholiday/vo/PaidHolidayInitVO.java
@@ -21,7 +21,7 @@ public class PaidHolidayInitVO {
     private String tphDyyyymmdd;
     private Integer seq;
     private Double investDaysSum;
-    private Integer investHours;
+    private Double investHours;
     private Double throughoutDaysSum;
     private Double throughoutHoursSum;
